# openapi.yaml
openapi: 3.0.3
info:
  title: Environmental Monitoring API
  version: 1.0.0
  description: |
    Production-grade REST API for environmental monitoring, mission management,
    visitor tracking, and geospatial analytics.

    Scenario-scoped routes are prefixed with:
    /api/v1/{scenario_name}/...
servers:
  - url: http://localhost:8000
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/{scenario_name}/measurement-types:
    get:
      tags: [Measurement Types]
      summary: List measurement types
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      responses:
        '200':
          description: Measurement types list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeasurementTypeResponse'
    post:
      tags: [Measurement Types]
      summary: Create measurement type
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementTypeCreate'
      responses:
        '201':
          description: Measurement type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementTypeResponse'
  /api/v1/{scenario_name}/instruments:
    get:
      tags: [Instruments]
      summary: List instruments
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      responses:
        '200':
          description: Instruments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstrumentResponse'
    post:
      tags: [Instruments]
      summary: Create instrument
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstrumentCreate'
      responses:
        '201':
          description: Instrument created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentResponse'
  /api/v1/{scenario_name}/instruments/{instrument_id}:
    patch:
      tags: [Instruments]
      summary: Update instrument
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstrumentUpdate'
      responses:
        '200':
          description: Instrument updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/{scenario_name}/measurements:
    get:
      tags: [Measurements]
      summary: List measurements with filtering and pagination
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: sort
          in: query
          schema:
            type: string
            default: timestamp
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
        - name: lat_min
          in: query
          schema:
            type: number
            format: double
        - name: lat_max
          in: query
          schema:
            type: number
            format: double
        - name: lon_min
          in: query
          schema:
            type: number
            format: double
        - name: lon_max
          in: query
          schema:
            type: number
            format: double
        - name: measurement_type_id
          in: query
          schema:
            type: string
            format: uuid
        - name: mission_id
          in: query
          schema:
            type: string
            format: uuid
        - name: label_ids
          in: query
          description: Comma-separated label UUIDs
          schema:
            type: string
      responses:
        '200':
          description: Paginated measurements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMeasurements'
    post:
      tags: [Measurements]
      summary: Create measurement
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementCreate'
      responses:
        '201':
          description: Measurement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
  /api/v1/{scenario_name}/measurements/{measurement_id}:
    get:
      tags: [Measurements]
      summary: Get measurement by id
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MeasurementId'
      responses:
        '200':
          description: Measurement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Measurements]
      summary: Update measurement
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MeasurementId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementUpdate'
      responses:
        '200':
          description: Measurement updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Measurements]
      summary: Delete measurement
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MeasurementId'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/{scenario_name}/measurements/{measurement_id}/labels:
    post:
      tags: [Measurements, Labels]
      summary: Assign label to measurement
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MeasurementId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeasurementLabelCreate'
      responses:
        '201':
          description: Label assignment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  measurement_id:
                    type: string
                    format: uuid
                  label_id:
                    type: string
                    format: uuid
                  reason:
                    type: string
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
  /api/v1/{scenario_name}/labels:
    get:
      tags: [Labels]
      summary: List labels
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      responses:
        '200':
          description: Labels list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LabelResponse'
    post:
      tags: [Labels]
      summary: Create label
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelCreate'
      responses:
        '201':
          description: Label created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelResponse'
...
components:
  parameters:
    ScenarioName:
      name: scenario_name
      in: path
      required: true
      schema:
        type: string
      description: Scenario identifier/name, e.g. drill-2026
    MeasurementId:
      name: measurement_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    MeasurementTypeCreate:
      type: object
      required: [name, unit]
      properties:
        name:
          type: string
          maxLength: 100
        unit:
          type: string
          maxLength: 50
    MeasurementTypeResponse:
      type: object
      required: [id, name, unit]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        unit: { type: string }
    InstrumentCreate:
      type: object
      required: [name, type]
      properties:
        name: { type: string, maxLength: 200 }
        type: { type: string, maxLength: 100 }
        calibration_date:
          type: string
          format: date-time
          nullable: true
    InstrumentUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 200 }
        type: { type: string, maxLength: 100 }
        calibration_date:
          type: string
          format: date-time
          nullable: true
    InstrumentResponse:
      type: object
      required: [id, name, type]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string }
        calibration_date:
          type: string
          format: date-time
          nullable: true
    MeasurementCreate:
      type: object
      required:
        [timestamp, location, measurement_type_id, value, unit, instrument_id]
      properties:
        timestamp: { type: string, format: date-time }
        location:
          $ref: '#/components/schemas/GeoJSONPoint'
        measurement_type_id: { type: string, format: uuid }
        value: { type: number, format: float }
        unit: { type: string, maxLength: 50 }
        instrument_id: { type: string, format: uuid }
        mission_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    MeasurementUpdate:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        location:
          $ref: '#/components/schemas/GeoJSONPoint'
        value: { type: number, format: float }
        unit: { type: string, maxLength: 50 }
        mission_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    MeasurementResponse:
      type: object
      required:
        [id, scenario_id, timestamp, measurement_type_id, value, unit, instrument_id, metadata, created_at]
      properties:
        id: { type: string, format: uuid }
        scenario_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        location:
          $ref: '#/components/schemas/GeoJSONPoint'
        measurement_type_id: { type: string, format: uuid }
        value: { type: number, format: float }
        unit: { type: string }
        instrument_id: { type: string, format: uuid }
        mission_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        measurement_type:
          $ref: '#/components/schemas/MeasurementTypeResponse'
        instrument:
          $ref: '#/components/schemas/InstrumentResponse'
    LabelCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100 }
        description:
          type: string
          nullable: true
          maxLength: 500
    LabelResponse:
      type: object
      required: [id, scenario_id, name]
      properties:
        id: { type: string, format: uuid }
        scenario_id: { type: string, format: uuid }
        name: { type: string }
        description:
          type: string
          nullable: true
    MeasurementLabelCreate:
      type: object
      required: [label_id]
      properties:
        label_id:
          type: string
          format: uuid
        reason:
          type: string
          nullable: true
          maxLength: 500
...
    GeoJSONPoint:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Point]
        coordinates:
          type: array
          minItems: 2
          maxItems: 3
          items:
            type: number
            format: double
...
    PaginatedMeasurements:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MeasurementResponse'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }
...